requirements.txt
beautifulsoup4
pandas
requests
psycopg2-binary
